Move helper types to `neon::types::function` submodule.
